Separate activity graph and  purchasers table into separate files, shifted all campaign progress related helpers into utility folder
