+++ src/components/utility/campaignProgressHelper.jsx
@@ -87,3 +87,56 @@ export const generatePaginationOptions = (indexAllPurchases) => {
   };
   return paginationOptions;
 };
+
+// Helper that generates yAxis Tick Values
+export const generateYAxisTickValues = (campaignStore) => {
+  const individualPurchaseCount = campaignStore.pastSevenDaysCount.map((dayCount) => dayCount.y);
+
+  const maxVal = Math.max(...individualPurchaseCount);
+  const minVal = Math.min(...individualPurchaseCount);
+  const purchaseCountRange = maxVal - minVal;
+
+  const yAxisInterval = Math.floor(purchaseCountRange / 5);
+  const yAxisTickValues = [];
+  // To ensure that second highest value is not too close to the top axis label
+  for (let i = 0; i < purchaseCountRange - yAxisInterval; i += yAxisInterval) {
+    yAxisTickValues.push(i);
+  }
+  // To ensure top of the graph always has the largest (label) value
+  yAxisTickValues.push(maxVal);
+
+  return yAxisTickValues;
+};
+
+// Helper that generates and pushes the past 7 days
+export const generatePastSevenDays = () => {
+  const pastSevenDaysArray = [];
+  for (let i = 7; i >= 0; i -= 1) {
+    // Start from today's date then progressive go backwards
+    const singleDate = new Date();
+    singleDate.setDate(singleDate.getDate() - i);
+
+    // Format into DD/MM
+    const options = { day: '2-digit', month: '2-digit' };
+    const formattedDate = singleDate.toLocaleDateString('en-GB', options);
+
+    // Append into an accumulative array
+    pastSevenDaysArray.push(formattedDate);
+  }
+  return pastSevenDaysArray;
+};
+
+// Helper that gets the lowest y-value
+export const getLowestYVal = (purchaseRange) => {
+  let yLow = null;
+  purchaseRange.forEach((day) => {
+    if (day.y < yLow || yLow === null) {
+      yLow = day.y;
+    }
+  });
+
+  if (yLow !== null) {
+    return yLow;
+  }
+  return 0;
+};
